semana10/sesion01/requirements.txt
backpack==0.1
blinker==1.4
cleo==0.6.8
Click==7.0
Faker==0.8.18
Flask==1.1.1
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
pyaml==16.12.2
Pygments==2.4.2
pylev==1.3.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
tzlocal==1.5.1
Werkzeug==0.16.0
wrapt==1.11.2
